Information About Datasets from TCGA Project

The checkTCGA function let's to check

DataSets: TCGA datasets' names for current release date and cohort.

Dates: TCGA datasets' dates of release.

checkTCGA(what, cancerType, date = NULL)

Arguments

what

One of DataSets or Dates.

cancerType

A character of length 1 containing abbreviation (Cohort code - http://gdac.broadinstitute.org/) of types of cancers to check for.

date

A NULL or character specifying from which date informations should be checked. By default (date = NULL) the newest available date is used. All available dates can be checked on http://gdac.broadinstitute.org/runs/ or by using checkTCGA('Dates') function. Required format 'YYYY-MM-DD'.

Value

If what='DataSets' a data.frame of available datasets' names (to pass to the downloadTCGA function) and sizes.

If what='Dates' a vector of available dates to pass to the downloadTCGA function.

Details

If what='DataSets' enables to check TCGA datasets' names for current release date and cohort.

If what='Dates' enables to check dates of TCGA datasets' releases.

Issues

If you have any problems, issues or think that something is missing or is not clear please post an issue on https://github.com/RTCGA/RTCGA/issues.

See also

RTCGA website http://rtcga.github.io/RTCGA/.

Other RTCGA: RTCGA-package, boxplotTCGA, convertTCGA, createTCGA, datasetsTCGA, downloadTCGA, expressionsTCGA, heatmapTCGA, infoTCGA, installTCGA, kmTCGA, mutationsTCGA, pcaTCGA, readTCGA, survivalTCGA, theme_RTCGA

Examples

############################# 

# names for current release date and cohort
checkTCGA('DataSets', 'BRCA')

#>    Size
#> 1   37M
#> 2   50K
#> 3  723K
#> 4  135K
#> 5  135K
#> 6   77K
#> 7   78K
#> 8  1.5M
#> 9   57K
#> 10 1.2K
#> 11 160K
#> 12 3.4M
#> 13  86K
#> 14  83M
#> 15 3.2G
#> 16 1.1M
#> 17  15M
#> 18 2.9M
#> 19  44M
#> 20 1.3M
#> 21 2.6G
#> 22 277M
#> 23 195M
#> 24 298M
#> 25  93M
#> 26 869M
#> 27 249M
#> 28 2.8G
#> 29 243M
#> 30  18M
#> 31  18M
#> 32 5.3M
#> 33 4.6M
#> 34  37M
#> 35 399M
#> 36  10M
#> 37 1.1G
#> 38  81M
#> 39 1.9M
#> 40  37M
#> 41 1.5G
#> 42 6.6M
#> 43 7.4M
#>                                                                                                                                             Name
#> 1  BRCA-FFPE.Merge_methylation__humanmethylation450__jhu_usc_edu__Level_3__within_bioassay_data_set_function__data.Level_3.2016012800.0.0.tar.gz
#> 2                   BRCA-FFPE.Merge_mirnaseq__illuminahiseq_mirnaseq__bcgsc_ca__Level_3__miR_gene_expression__data.Level_3.2016012800.0.0.tar.gz
#> 3                BRCA-FFPE.Merge_mirnaseq__illuminahiseq_mirnaseq__bcgsc_ca__Level_3__miR_isoform_expression__data.Level_3.2016012800.0.0.tar.gz
#> 4                         BRCA-FFPE.Merge_snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_hg18__seg.Level_3.2016012800.0.0.tar.gz
#> 5                         BRCA-FFPE.Merge_snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_hg19__seg.Level_3.2016012800.0.0.tar.gz
#> 6      BRCA-FFPE.Merge_snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_minus_germline_cnv_hg18__seg.Level_3.2016012800.0.0.tar.gz
#> 7      BRCA-FFPE.Merge_snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_minus_germline_cnv_hg19__seg.Level_3.2016012800.0.0.tar.gz
#> 8                                                                                 BRCA-FFPE.Methylation_Preprocess.Level_3.2016012800.0.0.tar.gz
#> 9                                                                               BRCA-FFPE.miRseq_Mature_Preprocess.Level_3.2016012800.0.0.tar.gz
#> 10                                                                                     BRCA-FFPE.miRseq_Preprocess.Level_3.2016012800.0.0.tar.gz
#> 11                                                                                        BRCA.Clinical_Pick_Tier1.Level_4.2016012800.0.0.tar.gz
#> 12                                                                                             BRCA.Merge_Clinical.Level_1.2016012800.0.0.tar.gz
#> 13                              BRCA.Merge_cna__illuminahiseq_dnaseqc__hms_harvard_edu__Level_3__segmentation__seg.Level_3.2016012800.0.0.tar.gz
#> 14       BRCA.Merge_methylation__humanmethylation27__jhu_usc_edu__Level_3__within_bioassay_data_set_function__data.Level_3.2016012800.0.0.tar.gz
#> 15      BRCA.Merge_methylation__humanmethylation450__jhu_usc_edu__Level_3__within_bioassay_data_set_function__data.Level_3.2016012800.0.0.tar.gz
#> 16                          BRCA.Merge_mirnaseq__illuminaga_mirnaseq__bcgsc_ca__Level_3__miR_gene_expression__data.Level_3.2016012800.0.0.tar.gz
#> 17                       BRCA.Merge_mirnaseq__illuminaga_mirnaseq__bcgsc_ca__Level_3__miR_isoform_expression__data.Level_3.2016012800.0.0.tar.gz
#> 18                       BRCA.Merge_mirnaseq__illuminahiseq_mirnaseq__bcgsc_ca__Level_3__miR_gene_expression__data.Level_3.2016012800.0.0.tar.gz
#> 19                    BRCA.Merge_mirnaseq__illuminahiseq_mirnaseq__bcgsc_ca__Level_3__miR_isoform_expression__data.Level_3.2016012800.0.0.tar.gz
#> 20                     BRCA.Merge_protein_exp__mda_rppa_core__mdanderson_org__Level_3__protein_normalization__data.Level_3.2016012800.0.0.tar.gz
#> 21                                BRCA.Merge_rnaseq__illuminahiseq_rnaseq__unc_edu__Level_3__exon_expression__data.Level_3.2016012800.0.0.tar.gz
#> 22                                BRCA.Merge_rnaseq__illuminahiseq_rnaseq__unc_edu__Level_3__gene_expression__data.Level_3.2016012800.0.0.tar.gz
#> 23                     BRCA.Merge_rnaseq__illuminahiseq_rnaseq__unc_edu__Level_3__splice_junction_expression__data.Level_3.2016012800.0.0.tar.gz
#> 24                                 BRCA.Merge_rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_genes__data.Level_3.2016012800.0.0.tar.gz
#> 25                      BRCA.Merge_rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_genes_normalized__data.Level_3.2016012800.0.0.tar.gz
#> 26                              BRCA.Merge_rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_isoforms__data.Level_3.2016012800.0.0.tar.gz
#> 27                   BRCA.Merge_rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_isoforms_normalized__data.Level_3.2016012800.0.0.tar.gz
#> 28                        BRCA.Merge_rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__exon_quantification__data.Level_3.2016012800.0.0.tar.gz
#> 29                    BRCA.Merge_rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__junction_quantification__data.Level_3.2016012800.0.0.tar.gz
#> 30                             BRCA.Merge_snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_hg18__seg.Level_3.2016012800.0.0.tar.gz
#> 31                             BRCA.Merge_snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_hg19__seg.Level_3.2016012800.0.0.tar.gz
#> 32          BRCA.Merge_snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_minus_germline_cnv_hg18__seg.Level_3.2016012800.0.0.tar.gz
#> 33          BRCA.Merge_snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_minus_germline_cnv_hg19__seg.Level_3.2016012800.0.0.tar.gz
#> 34       BRCA.Merge_transcriptome__agilentg4502a_07_3__unc_edu__Level_3__unc_lowess_normalization_gene_level__data.Level_3.2016012800.0.0.tar.gz
#> 35                                                                                     BRCA.Methylation_Preprocess.Level_3.2016012800.0.0.tar.gz
#> 36                                                                                    BRCA.Mutation_Packager_Calls.Level_3.2016012800.0.0.tar.gz
#> 37                                                                                 BRCA.Mutation_Packager_Coverage.Level_3.2016012800.0.0.tar.gz
#> 38                                                                          BRCA.Mutation_Packager_Oncotated_Calls.Level_3.2016012800.0.0.tar.gz
#> 39                                                                                      BRCA.RPPA_AnnotateWithGene.Level_3.2016012800.0.0.tar.gz
#> 40                                                                                     BRCA.mRNA_Preprocess_Median.Level_3.2016012800.0.0.tar.gz
#> 41                                                                                         BRCA.mRNAseq_Preprocess.Level_3.2016012800.0.0.tar.gz
#> 42                                                                                   BRCA.miRseq_Mature_Preprocess.Level_3.2016012800.0.0.tar.gz
#> 43                                                                                          BRCA.miRseq_Preprocess.Level_3.2016012800.0.0.tar.gz

## Not run: ------------------------------------
# checkTCGA('DataSets', 'OV', tail(checkTCGA('Dates'))[3])
# #checkTCGA('DataSets', 'OV', checkTCGA('Dates')[5]) # error
## ---------------------------------------------
# dates of TCGA datasets' releases.
checkTCGA('Dates')

#>  [1] "2011-10-26" "2011-11-15" "2011-11-28" "2011-12-06" "2011-12-30"
#>  [6] "2012-01-10" "2012-01-24" "2012-02-17" "2012-03-06" "2012-03-21"
#> [11] "2012-04-12" "2012-04-25" "2012-05-15" "2012-05-25" "2012-06-06"
#> [16] "2012-06-23" "2012-07-07" "2012-07-25" "2012-08-04" "2012-08-25"
#> [21] "2012-09-13" "2012-10-04" "2012-10-18" "2012-10-20" "2012-10-24"
#> [26] "2012-11-02" "2012-11-14" "2012-12-06" "2012-12-21" "2013-01-16"
#> [31] "2013-02-03" "2013-02-22" "2013-03-09" "2013-03-26" "2013-04-06"
#> [36] "2013-04-21" "2013-05-08" "2013-05-23" "2013-06-06" "2013-06-23"
#> [41] "2013-07-15" "2013-08-09" "2013-09-23" "2013-10-10" "2013-11-14"
#> [46] "2013-12-10" "2014-01-15" "2014-02-15" "2014-03-16" "2014-04-16"
#> [51] "2014-05-18" "2014-06-14" "2014-07-15" "2014-09-02" "2014-10-17"
#> [56] "2014-12-06" "2015-02-02" "2015-02-04" "2015-04-02" "2015-06-01"
#> [61] "2015-08-21" "2015-11-01" "2016-01-28"

############################# 
## Not run: ------------------------------------
# # TCGA datasets' names availability for 
# # current release date and cancer type.
# 
# releaseDate <- '2015-08-21'
# cancerTypes <- c('OV', 'BRCA')
# 
# cancerTypes %>% sapply(function(element){
#   grep(x = checkTCGA('DataSets', element, releaseDate)[, 1], 
#       pattern = 'humanmethylation450', value = TRUE) %>%
#        as.vector()
#        })
#        
## ---------------------------------------------