Create RTCGA.dataType.releaseDate-like Packages with TCGA Datasets

This uber functionality allows to create packages with datasets from the TCGA project. Example of such a package created with createTCGA is https://github.com/RTCGA/RTCGA.clinical.20160128. Resulted package is a package of type http://www.bioconductor.org/packages/3.4/bioc/vignettes/ExperimentHubData/inst/doc/ExperimentHubData.html, it consists of regular structure like R and man directories and also cosists of a vignette and README.md file where the whole construction process is decribed. The only part which must be pre-specified by a user is a DESCRIPTION file with additional extra records: TCGAdataSet, TCGAreleaseDate (releaseDate and dataSet to be used in downloadTCGA) and TCGAdataType (a type of data to be used in readTCGA).

createTCGA(description = file.path(getwd(), "DESCRIPTION"),
  cohorts = sub("-counts", "", rownames(infoTCGA())), tempDir = "data_tmp",
  clean = FALSE)

Arguments

description

A regular http://r-pkgs.had.co.nz/description.html file with additional extra records: TCGAdataSet, TCGAreleaseDate (releaseDate and dataSet to be used in downloadTCGA) and TCGAdataType (a type of data to be used in readTCGA).

cohorts

Names of cancer type cohorts to be included in the package. For possible names check infoTCGA.

tempDir

The name of temporary directory in which the data will be stored during package creation.

clean

A logical: should the tempDir directory be removed after package creation.

Value

A fully equpied package, ready to be shipped on Bioconductor's ExperimentHub.

Issues

If you have any problems, issues or think that something is missing or is not clear please post an issue on https://github.com/RTCGA/RTCGA/issues.

See also

RTCGA website http://rtcga.github.io/RTCGA/.

Other RTCGA: RTCGA-package, boxplotTCGA, checkTCGA, convertTCGA, datasetsTCGA, downloadTCGA, expressionsTCGA, heatmapTCGA, infoTCGA, installTCGA, kmTCGA, mutationsTCGA, pcaTCGA, readTCGA, survivalTCGA, theme_RTCGA

Examples

## Not run: ------------------------------------
# 
# # https://github.com/RTCGA/RTCGA.clinical.20160128
# # RTCGA.clinical.20160128 was created with
# cat(
#  readLines(
#    'https://raw.githubusercontent.com/RTCGA/RTCGA.clinical.20160128/master/DESCRIPTION'
# ), file = 'DESC', sep = '\n')
# createTCGA('DESC', clean = TRUE)
# 
## ---------------------------------------------