RTCGA - The Family of R Packages with Data from The Cancer Genome Atlas Study

Snapshots of the clinical, mutations, CNVs, rnaseq, RPPA, mRNA, miRNASeq and methylation datasets from the 2016-01-28 release date (check all dates of release with checkTCGA('Dates')) are included in the RTCGA family (factory) that contains below packages:

RTCGA.rnaseq.20160128 rnaseq.20160128

RTCGA.clinical.20160128 clinical.20160128

RTCGA.mutations.20160128 mutations.20160128

RTCGA.CNV.20160128 CNV.20160128

RTCGA.RPPA.20160128 RPPA.20160128

RTCGA.mRNA.20160128 mRNA.20160128

RTCGA.miRNASeq.20160128 miRNASeq.20160128

RTCGA.methylation.20160128 methylation.20160128

Snapshots of the clinical, mutations, CNVs, rnaseq, RPPA, mRNA, miRNASeq and methylation datasets from the 2015-11-01 release date (check all dates of release with checkTCGA('Dates')) are also included in the RTCGA family (factory).

RTCGA.rnaseq rnaseq

RTCGA.clinical clinical

RTCGA.mutations mutations

RTCGA.CNV CNV

RTCGA.RPPA RPPA

RTCGA.mRNA mRNA

RTCGA.miRNASeq miRNASeq

RTCGA.methylation methylation

RTCGA.PANCAN12 (not from TCGA)

Details

For more detailed information visit RTCGA family website https://rtcga.github.io/RTCGA. One can install all data packages with installTCGA.

Issues

If you have any problems, issues or think that something is missing or is not clear please post an issue on https://github.com/RTCGA/RTCGA/issues.

See also

RTCGA website http://rtcga.github.io/RTCGA.

Other RTCGA: RTCGA-package, boxplotTCGA, checkTCGA, convertTCGA, createTCGA, downloadTCGA, expressionsTCGA, heatmapTCGA, infoTCGA, installTCGA, kmTCGA, mutationsTCGA, pcaTCGA, readTCGA, survivalTCGA, theme_RTCGA

Examples

# installation of packages containing snapshots
# of TCGA project's datasets

## Not run: ------------------------------------
# 
# ## RTCGA GitHub development newest versions
# library(RTCGA)
# ?installTCGA
# 
# ## Bioconductor releases for data from 2016-01-28 release
# source('http://bioconductor.org/biocLite.R')
# biocLite(RTCGA.clinical.20160128)
# biocLite(RTCGA.mutations.20160128)
# biocLite(RTCGA.rnaseq.20160128)
# biocLite(RTCGA.CNV.20160128)
# biocLite(RTCGA.RPPA.20160128)
# biocLite(RTCGA.mRNA.20160128)
# biocLite(RTCGA.miRNASeq.20160128)
# biocLite(RTCGA.methylation.20160128)
# 
# ## Bioconductor releases for data from 2015-11-01 release
# source('http://bioconductor.org/biocLite.R')
# biocLite(RTCGA.clinical)
# biocLite(RTCGA.mutations)
# biocLite(RTCGA.rnaseq)
# biocLite(RTCGA.CNV)
# biocLite(RTCGA.RPPA)
# biocLite(RTCGA.mRNA)
# biocLite(RTCGA.miRNASeq)
# biocLite(RTCGA.methylation)
# 
# # use cases and examples + more data info
# browseVignettes('RTCGA')
## ---------------------------------------------