Install Data Packages from RTCGA Family

Function installs data packages from https://github.com/RTCGA/. Packages are listed in datasetsTCGA.

installTCGA(packages = c("RTCGA.clinical.20160128",
  "RTCGA.mutations.20160128", "RTCGA.rnaseq.20160128", "RTCGA.RPPA.20160128",
  "RTCGA.mRNA.20160128", "RTCGA.CNV.20160128", "RTCGA.miRNASeq.20160128",
  "RTCGA.PANCAN12.20160128", "RTCGA.methylation.20160128"),
  build_vignettes = TRUE, ...)

Arguments

packages

A character specifing the names of the data packages to be installed. By default installs all packages from .20160128 release.

build_vignettes

Should vignettes be build.

...

Further arguments passed to install_github.

Issues

If you have any problems, issues or think that something is missing or is not clear please post an issue on https://github.com/RTCGA/RTCGA/issues.

See also

RTCGA website http://rtcga.github.io/RTCGA.

Other RTCGA: RTCGA-package, boxplotTCGA, checkTCGA, convertTCGA, createTCGA, datasetsTCGA, downloadTCGA, expressionsTCGA, heatmapTCGA, infoTCGA, kmTCGA, mutationsTCGA, pcaTCGA, readTCGA, survivalTCGA, theme_RTCGA

Examples

## Not run: ------------------------------------
# installTCGA() # it installs all!!! of them
# installTCGA('RTCGA.clinical.20160128')
## ---------------------------------------------